chore: make Base for the IT tests
