+++ src/it/java/dev/jbang/it/CommandResultAssert.java
@@ -79,4 +79,12 @@ public class CommandResultAssert extends AbstractAssert<CommandResultAssert, Com
 		}
 		return this;
 	}
+
+    public CommandResultAssert outNotContains(String string) {
+        isNotNull();
+        if (actual.out.contains(string)) {
+            failWithMessage("Expected output to not contain <%s> but was <%s>", string, actual.out);
+        }
+        return this;
+    }
 }
